cross entropy without ignoreindex option

+++ datasets/weedcluster.py
@@ -34,7 +34,7 @@ class WeedClusterDataset(Dataset):
     def __getitem__(self, index):
         img = torch.from_numpy(np.array(Image.open(self.root + self.split +  '/images/rgb/' + self.images[index]).convert('RGB')).reshape((3,512,512)))
         # print("img shape: ", img.shape)
-        target = torch.from_numpy(np.int(np.array(Image.open(self.root + self.split +  '/labels/weed_cluster/' + self.images[index][:-4] + '.png'))/255))
+        target = torch.from_numpy((np.array(Image.open(self.root + self.split +  '/labels/weed_cluster/' + self.images[index][:-4] + '.png'))/255).astype(int))
         # print("target shape: ", target.shape)
         # print("target: ", target)
